incremental checkin
git-svn-id: file:///local/rexx/oorexx/oorexx.svn@2284 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ sandbox/rick/init/platform/windows/orxscrpt/orxidispatch.cpp
@@ -869,7 +869,7 @@ STDMETHODIMP OrxScript::Invoke(DISPID pDispID, REFIID riid, LCID plcid,
    */
   if(SUCCEEDED(RetCode)) {
     // Now convert it to wide characters.
-    ISMaxLen = strlen(FInvokeString)+1;
+    ISMaxLen = (int)strlen(FInvokeString)+1;
     //   >>> ??? <<< Code for the case where this has no invocation string.
     if(ISMaxLen > 1) {
       if(ISMaxLen > (sizeof(invokeString)/sizeof(OLECHAR))) {
